requirements.txt
streamlit
langchain
langchain_community
zhconv
langchain_ollama
tiktoken
sentence-transformers
torch